DocumentStorage/Services/CloudDocumentStorage.cs
using DocumentStorage.Models;
using Newtonsoft.Json.Linq;

namespace DocumentStorage.Services
{
    public class CloudDocumentStorage : ICloudDocumentStorage
    {
        private readonly IConfiguration _configuration;
        private readonly ILogger<CloudDocumentStorage> _logger;

        public CloudDocumentStorage(IConfiguration configuration, ILogger<CloudDocumentStorage> logger)
        {
            _configuration = configuration;
            _logger = logger;
        }

        public async Task<Document> GetDocumentAsync(string id)
        {
            _logger.LogInformation($"Getting document {id} from cloud storage.");
            // Replace with actual cloud storage API calls
            // Example: var document = await CloudStorageService.GetDocumentAsync(id);

            // For simplicity, return a sample document
            return new Document 
            { 
                Id = id, 
                Tags = new List<string> { "cloud", "document" }, 
                Data = new JObject() 
            };
        }

        public async Task<Document> StoreDocumentAsync(Document document)
        {
            // Implement code to store document in cloud storage
            _logger.LogInformation($"Storing document {document.Id} in cloud storage.");
            // Replace with actual cloud storage API calls
            // Example: await CloudStorageService.StoreDocumentAsync(document);

            // For simplicity, return the same document
            return document;
        }

        public async Task<Document> UpdateDocumentAsync(string id, Document document)
        {
            // Implement code to update document in cloud storage
            _logger.LogInformation($"Updating document {id} in cloud storage.");
            // Replace with actual cloud storage API calls
            // Example: await CloudStorageService.UpdateDocumentAsync(id, document);

            // For simplicity, return the updated document
            return document;
        }
    }

}
